# public/swagger.yaml
openapi: "3.0.0"
info:
  title: Books API
  description: A simple API to manage books
  version: "1.0.0"

servers:
  - url: http://localhost:9000

paths:
  /books:
    get:
      summary: Get all books
      operationId: getBooks
      responses:
        '200':
          description: A list of books

    post:
      summary: Add a new book
      operationId: addBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - isbn
                - title
                - copyrightYear
                - status
              properties:
                isbn:
                  type: string
                title:
                  type: string
                subtitle:
                  type: string
                copyrightYear:
                  type: integer
                status:
                  type: string
      responses:
        '201':
          description: Book created
        '400':
          description: Bad Request

  /books/{id}:
    get:
      summary: Get a book by ID
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single book
        '404':
          description: Not found

    put:
      summary: Update a book
      operationId: updateBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isbn:
                  type: string
                title:
                  type: string
                subtitle:
                  type: string
                copyrightYear:
                  type: integer
                status:
                  type: string
      responses:
        '200':
          description: Book updated
        '404':
          description: Not found

    delete:
      summary: Delete a book
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Book deleted
        '404':
          description: Not found